delete student jwt

+++ routes/student.js
@@ -42,26 +42,16 @@ router.put("/:id", verify, async (req, res) => {
 
 //delete original
 
-// router.delete("/:id", verify, async (req, res) => {
-//   if (req.user.id === req.params.id || req.user.isAdmin) {
-//     try {
-//       await Student.findByIdAndDelete(req.params.id);
-//       res.status(200).json("User has been deleted");
-//     } catch (err) {
-//       res.status(500).json(err);
-//     }
-//   } else {
-//     res.status(403).json("you can delete your own account");
-//   }
-// });
-
-//delete test case
-router.delete("/:id", async (req, res) => {
-  try {
-    await Student.findByIdAndDelete(req.params.id);
-    res.status(200).json("User has been deleted");
-  } catch (err) {
-    res.status(500).json(err);
+router.delete("/:id", verify, async (req, res) => {
+  if (req.user.id === req.params.id || req.user.isAdmin) {
+    try {
+      await Student.findByIdAndDelete(req.params.id);
+      res.status(200).json("User has been deleted");
+    } catch (err) {
+      res.status(500).json(err);
+    }
+  } else {
+    res.status(403).json("you can delete your own account");
   }
 });
